+++ docs/CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 ## [Unreleased]
 
+### Added
+
+- Added a command to clone the repos from the config
 
 
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/